getDatabaseFiles moved to common
