+++ server/lib/tfs-tfvc.js
@@ -331,30 +331,7 @@ const downloadDatabaseScript = (changesetId, databaseName, scripts) => {
  * Get all database scripts.
  */
 const getDatabaseData = (changesetId, files) => {
-  const databases = {};
-
-  _.filter(files, f => common.isDatabaseConnection(f.path)).forEach(file => {
-    const script = common.getDatabaseScriptDetails(file.path);
-    const settings = common.getDatabaseSettingsDetails(file.path);
-
-    if (script) {
-      databases[script.database] = databases[script.database] || [];
-      databases[script.database].push({
-        ...script,
-        id: file.id,
-        path: file.path
-      });
-    }
-
-    if (settings) {
-      databases[settings.database] = databases[settings.database] || [];
-      databases[settings.database].push({
-        ...settings,
-        id: file.id,
-        path: file.path
-      });
-    }
-  });
+  const databases = common.getDatabaseFiles(files);
 
   return Promise.map(Object.keys(databases), (databaseName) => downloadDatabaseScript(changesetId, databaseName, databases[databaseName]), { concurrency: 2 });
 };
